// frontend/src/gen/schemas/getApiV1IamPermission200PermissionsItem.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * Bones
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiV1IamPermission200PermissionsItemServiceType } from "./getApiV1IamPermission200PermissionsItemServiceType";
import type { GetApiV1IamPermission200PermissionsItemUsersItem } from "./getApiV1IamPermission200PermissionsItemUsersItem";

export type GetApiV1IamPermission200PermissionsItem = {
    createdAt: string;
    isServiceWide: boolean;
    /** @maxLength 255 */
    name: string;
    permissionId: string;
    readAll: boolean;
    serviceType: GetApiV1IamPermission200PermissionsItemServiceType;
    updatedAt: string;
    users: GetApiV1IamPermission200PermissionsItemUsersItem[];
    writeAll: boolean;
};